Migrate org.mockito.Matchers#any* to org.mockito.ArgumentMatchers
The former is deprecated and replaced by the latter in Mockito 2. However, there is a
functional difference: ArgumentMatchers will reject `null` and check the type
if the matcher specified a type (e.g. `any(Class)` or `anyInt()`). `any()` will
remain to accept anything.

For more information see go/matchers-any-null-lsc

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:250287304:BASE:250411780:1559100945691:2daa4cb5
PiperOrigin-RevId: 250458606
Change-Id: Ic046c51d6a35be6dd2417f48b1351ad6f2b5f7f7